cleanup and refactor rect-pred pipeline
